Move the route query to main
Trying to use the manager client before the manager was started lead to
an error because the client cache was not ready yet.

```
2024-02-09T20:18:07Z	ERROR	setup	unable to create controller	{"controller": "ImageClusterInstall", "error": "failed to create base URL: the cache is not started, can not read objects"}
```

To avoid this, create an uncached client in main and use that only for
querying the route base URL.
